Add gc skeleton code
